tiny changes of values

+++ src/main/java/net/akami/mistream/play/list/ForwardMovement.java
@@ -8,7 +8,7 @@ public class ForwardMovement extends TerminalSequence {
 
     private float speed;
 
-    public ForwardMovement(int time, float speed) {
+    public ForwardMovement(float time, float speed) {
         super(time, null);
         this.speed = speed;
     }
